+++ src/lib/ares_parse_naptr_reply.c
@@ -37,13 +37,13 @@
 #include "ares_private.h"
 
 int
-ares_parse_naptr_reply (const unsigned char *abuf, int alen,
+ares_parse_naptr_reply (const unsigned char *abuf, size_t alen,
                         struct ares_naptr_reply **naptr_out)
 {
   unsigned int qdcount, ancount, i;
   const unsigned char *aptr, *vptr;
   int status, rr_type, rr_class, rr_len;
-  long len;
+  size_t len;
   char *hostname = NULL, *rr_name = NULL;
   struct ares_naptr_reply *naptr_head = NULL;
   struct ares_naptr_reply *naptr_last = NULL;
