+++ src/lib/ares_send.c
@@ -28,7 +28,7 @@
 #include "ares_dns.h"
 #include "ares_private.h"
 
-void ares_send(ares_channel channel, const unsigned char *qbuf, int qlen,
+void ares_send(ares_channel channel, const unsigned char *qbuf, size_t qlen,
                ares_callback callback, void *arg)
 {
   struct query *query;
